Update and rename Calculator_1.5.0.cpp to Calculator_1.5.1.cpp
I've completed the while loop I mentioned before, but for such a simple code I got up to 200 lines of code, the reason for this is that instead of assigning a function, I'm copy and pasting each code over and over again. In update 2.0 I will assign a function and shorten the code as much as possible.